refactor(sourcepos): extract shared `parse_sourcepos` helper
Move `parse_sourcepos` (and the `Sourcepos` struct) into
`helpers::sourcepos` and reuse it from both `fix_img` and `fix_link`,
removing the inline duplicate in `fix_link`.

Minor behavior change in `fix_link`: when `data-sourcepos` is present
but malformed, the warning is now emitted without positions (via the
no-sourcepos branch) instead of being silently dropped.
